remove(40-define-functions): delete 320-parameters lesson, move truncate exercise to 150-return
Урок дублировал введение параметров из 100-define-function. По схеме Python:
упражнение truncate переехало в 150-return (в Python оно живёт именно там),
уникальные идеи (произвольность имени параметра, важность порядка) влиты
в 100-define-function. es-дубль description.es.yml урока 150 синхронизирован.
