+++ modules/40-define-functions/100-define-function/ru/README.md
@@ -83,6 +83,31 @@ printAverage(6, 4); // => 5
 
 Здесь `a` и `b` — входные параметры, `total` содержит их сумму, `average` получается делением суммы на 2, а `console.log()` выводит результат.
 
+## Имена и порядок параметров
+
+Имя параметра может быть любым — главное, чтобы оно отражало смысл значения, которое попадёт внутрь. С внешним кодом оно никак не связано:
+
+```javascript
+function getLastChar(str) {
+  return str[str.length - 1];
+}
+
+// Внутри функции str будет равна 'Winter is coming'.
+// Имя переменной снаружи не связано с именем параметра
+const text = 'Winter is coming';
+console.log(getLastChar(text)); // => g
+```
+
+Когда параметров два и больше, для большинства функций становится важен порядок их передачи. Если его поменять, функция отработает по-другому:
+
+```javascript
+// Первый параметр — что ищем, второй — на что меняем
+console.log('google'.replace('go', 'mo')); // => moogle
+
+// Ничего не заменилось: внутри 'google' нет 'mo'
+console.log('google'.replace('mo', 'go')); // => google
+```
+
 В JavaScript есть и другая, сокращённая форма записи функций — стрелочная. Ей посвящён отдельный урок дальше в этом модуле.
 
 ## Переиспользование и читаемость
