+++ modules/40-define-functions/150-return/index.js
@@ -1,9 +1,8 @@
-// BEGIN
-function sayHurray() {
-  return 'hurray! hurray! hurray!';
+function truncate(text, length) {
+  // BEGIN
+  const result = `${text.slice(0, length)}...`;
+  return result;
+  // END
 }
-// END
 
-console.log(sayHurray());
-
-export default sayHurray;
+export default truncate;
