+++ modules/40-define-functions/150-return/test.js
@@ -4,5 +4,6 @@ import { expect, test } from 'vitest';
 import f from './index.js';
 
 test('test', () => {
-  expect(f()).toBe('hurray! hurray! hurray!');
+  expect(f('текст', 3)).toBe('тек...');
+  expect(f('и пошла вода', 5)).toBe('и пош...');
 });
