--- modules/40-define-functions/320-define-functions-parameters/Makefile
@@ -1,2 +1,0 @@
-test:
-	@ test.sh
