--- modules/40-define-functions/320-define-functions-parameters/en/data.yml
@@ -1,2 +1,0 @@
----
-name: Function parameters
