--- modules/40-define-functions/320-define-functions-parameters/es/data.yml
@@ -1,6 +1,0 @@
----
-name: Parámetros de funciones
-tips:
-  - >
-    [Parámetros de
-    función](https://es.m.wikipedia.org/wiki/Argumento_(inform%C3%A1tica))
