--- modules/40-define-functions/320-define-functions-parameters/index.js
@@ -1,8 +1,0 @@
-function truncate(text, length) {
-  // BEGIN
-  const result = `${text.slice(0, length)}...`;
-  return result;
-  // END
-}
-
-export default truncate;
