--- modules/40-define-functions/320-define-functions-parameters/ru/data.yml
@@ -1,6 +1,0 @@
----
-name: Параметры функций
-tips:
-  - >
-    [Параметры
-    функции](https://ru.wikipedia.org/wiki/Параметр_(программирование))
