--- modules/40-define-functions/320-define-functions-parameters/test.js
@@ -1,9 +1,0 @@
-// @ts-check
-
-import { expect, test } from 'vitest';
-import f from './index.js';
-
-test('test', () => {
-  expect(f('текст', 3)).toBe('тек...');
-  expect(f('и пошла вода', 5)).toBe('и пош...');
-});
